servers/models/requirements.txt
newsapi-python
wikipedia-api
flask
flask-cors
langchain
langchain-core